requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3

# Database
psycopg2-binary>=2.9  # if using PostgreSQL (remove if only SQLite)
sqlparse>=0.4.4

# AI + LangChain
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2
langchain-chroma>=0.1
chromadb>=0.5
sentence-transformers>=2.2
huggingface-hub>=0.23

# Ollama integration
ollama>=0.1.6

# Utilities
python-dotenv>=1.0
requests>=2.32
